Domain 업로드 ( 기본키 Long형으로 변환 )

+++ src/main/java/com/devhive03/Model/DAO/CommunityPostLikesList.java
@@ -9,7 +9,7 @@ public class CommunityPostLikesList {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
-    private Integer likeCommunityPostID;
+    private Long likeCommunityPostID;
 
     @ManyToOne
     @JoinColumn(name = "like_user_id", nullable = false)
